Show email content on iframe DOMContentLoaded
This is a lot faster than listening to on load, which waits for
everything including the images.

While it is not possible to listen to DOMContentLoaded on the iframe
directly, firing a callback from within the iframe via window.parent
does the job just as well in this same-domain case.