Replicate Sass variables using CSS variables.
Setup CSS variables to mimic most of the Sass variables that were
previously in use by the Sass stylesheets.

This introduces some additional variables to capture the various
variations on colors that were based on Sass `lighten()` or `darken()`
calls on the other variables.

The one exception that is no longer a variable is the
`$rep-grid-breakpoint-sm: 768px`, since CSS media queries can't use CSS
variables.