Use plain CSS to eliminate need for Sprockets and Sass building (#93)
* Use plain CSS to eliminate need for Sprockets and Sass building

- Compile the Sass files into resulting plain CSS and use those
  pre-built files so that no Sass build step is needed.
- Inline the CSS and favicon into the default layout file to simplify
  compatibility across various different asset solutions. This way it
  doesn't matter if the app is using Asset Pipeline/Sprockets,
  Webpacker, Vite, or none of the above.
- Remove Turbolinks javascript, since the library is deprecated.
  Embedding applications could choose to add it if desired, but since
  the functionality was optional, it seems simplest to skip including it
  directly in the gem.

* Replicate Sass variables using CSS variables.

Setup CSS variables to mimic most of the Sass variables that were
previously in use by the Sass stylesheets.

This introduces some additional variables to capture the various
variations on colors that were based on Sass `lighten()` or `darken()`
calls on the other variables.

The one exception that is no longer a variable is the
`$rep-grid-breakpoint-sm: 768px`, since CSS media queries can't use CSS
variables.